backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
requests
python-dotenv
